Moved 'parser_trim_buffers' test to its proper place
